requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Script
Flask-WTF
psycopg2
werkzeug
